refactor: ゲームデータ同期の実装部分を別コンポーネントに移動
